Split up AtomType and MoleculeType into separate headers
