+++ src/gromacs/nblib/molecules.cpp
@@ -6,25 +6,6 @@
 
 namespace nblib {
 
-AtomType::AtomType()
-: atomName_(""),
-  mass_(0),
-  charge_(0),
-  c6_(0),
-  c12_(0)
-{}
-
-AtomType::AtomType(std::string atomName, real mass, real charge, real c6, real c12)
-: atomName_(std::move(atomName)),
-  mass_(mass),
-  charge_(charge),
-  c6_(c6),
-  c12_(c12)
-{}
-
-std::string AtomType::getName() const { return atomName_; }
-
-
 MoleculeType::MoleculeType(std::string moleculeName) : moleculeName_(std::move(moleculeName)) {}
 
 MoleculeType& MoleculeType::addAtom(const std::string &moleculeAtomName, const std::string &residueName, AtomType const &atom)
